Clarify the LLVM dependency on Windows
I had some trouble understanding that I had to install LLVM separately and it wasn't included with the Visual Studio as a dependency. I'm creating this PR because it confused me and I saw some people on the discord go through this as well, and that's how I "figured it out myself".